… man/avocado.rst