With many tests producing quite long logs my Jenkins instance is failing to publish the results due to too-big-xunit files. This PR adds support to limit the results size by embedding only partial logs in case they grew too big. The value is per-test so users need to adjust the limits according to number of tests they execute. Anyway calculating it from number of tests would be IMO confusing.