@ldoktor 2018-04-11
Btw the travis issue is fixable by:

```diff
diff --git a/selftests/unit/test_xunit.py b/selftests/unit/test_xunit.py
index cda9d8dc..4267f1c3 100644
--- a/selftests/unit/test_xunit.py
+++ b/selftests/unit/test_xunit.py
@@ -96,7 +96,7 @@ class xUnitSucceedTest(unittest.TestCase):
 
     def test_max_test_log_size(self):
         log = tempfile.NamedTemporaryFile(dir=self.tmpdir, delete=False)
-        log_content = "1234567890" * 100
+        log_content = b"1234567890" * 100
         log.write(log_content)
         log_path = log.name
         log.close()
```

(I'd force-update, but Amador already made some comments...)