@apahim 2018-04-13
Just not to trash my work on the `DataSize` optimization, let me record here the diff and the achieved results:

```diff
diff --git a/avocado/utils/data_structures.py b/avocado/utils/data_structures.py
index 31f9166a..1465816c 100644
--- a/avocado/utils/data_structures.py
+++ b/avocado/utils/data_structures.py
@@ -22,7 +22,6 @@ avocado core code or plugins.
 """
 
 
-import re
 import sys
 import math
 
@@ -283,28 +282,54 @@ class DataSize(object):
     :type data: str
     """
 
-    def __init__(self, data):
-        pattern = r"^(\d+)([bkmgt])?$"  # Number and optional string
-        match = re.match(pattern, data, re.IGNORECASE)
-
-        if match is None:
+    def __init__(self, size):
+        try:
+            norm_size = size.strip().lower()
+            last = norm_size[-1]
+            if last.isdigit():
+                self._value = int(norm_size)
+                self._unit = 'b'
+            elif last in SIZE_MULTIPLIERS:
+                self._value = int(norm_size[:-1])
+                self._unit = last
+            else:
+                raise ValueError("Unable to convert '%s'" % size)
+        except ValueError:
             raise InvalidDataSize('String not in size+unit format (i.e. '
                                   '"10M", "100k", ...)')
 
-        self._value, unit = match.groups()
-        if unit is None:
-            self._unit = 'b'
-        else:
-            self._unit = unit.lower()
-
     @property
     def value(self):
-        return int(self._value)
+        return self._value
 
     @property
     def unit(self):
         return self._unit
 
+    @property
+    def b(self):
+        return int(self._value * SIZE_MULTIPLIERS[self._unit])
+
+    @property
+    def k(self):
+        return int(self._value * SIZE_MULTIPLIERS[self._unit] /
+                   SIZE_MULTIPLIERS['k'])
+
+    @property
+    def m(self):
+        return int(self._value * SIZE_MULTIPLIERS[self._unit] /
+                   SIZE_MULTIPLIERS['m'])
+
+    @property
+    def g(self):
+        return int(self._value * SIZE_MULTIPLIERS[self._unit] /
+                   SIZE_MULTIPLIERS['g'])
+
+    @property
+    def t(self):
+        return int(self._value * SIZE_MULTIPLIERS[self._unit] /
+                   SIZE_MULTIPLIERS['t'])
+
     def __getattr__(self, attr):
         """
         Creates one extra attribute per available conversion unit,
```
**RESULTS**

Test code:
```python
import timeit

number = 1000000

t1 = timeit.timeit('data_structures.DataSize("1M").b',
                   setup='from avocado.utils import data_structures',
                   number=number)
t2 = timeit.timeit('data_structures.size_to_bytes("1M")',
                   setup='from avocado.utils import data_structures',
                   number=number)

print("%s x DataSize('1M').b: %ss" % (number, t1))
print("%s x size_to_bytest('1M'): %ss" % (number,t2))

```

Before the patch:
```
1000000 x DataSize('1M').b: 2.68798303604s
1000000 x size_to_bytest('1M'): 0.646973133087s
```

After the patch:
```
1000000 x DataSize('1M').b: 1.13909196854s
1000000 x size_to_bytest('1M'): 0.64511013031s
```

The optimizations are covering:
- Faster parsing: dropping the `re.match()`.
- Define the known properties instead of relying solely on `__getattr__`, keeping the `__getattr__` to make sure new properties added to `SIZE_MULTIPLIERS` will be automatically available.
- Access value and unit using the attributes instead of using the properties.

The overhead that still exists now is consequence of the `Class` handling when compared to the function handling. Further optimizations should consider using a memoize decorator in the properties, to perform better in the case we are using to test the code (millions of calls). But since that doesn't seem to be a common use case and as it would not improve the performance for one-time calls, I don't see myself working on that.

Cheers!