@ldoktor 2018-04-13
@apahim I don't want to be rude, but if you take a look at my comments I suggested adding the methods rather than dynamically assign them ;-) Anyway I'm definitely for including your optimizations so please send a PR.