@ldoktor 2018-04-13
@apahim @clebergnu: Please let me know what should I do with this PR. I'd like to have the `size_to_bytes` but if you're strongly against, I can replace it with `DataSize`. The rest seems OK, right?