@apahim 2018-04-13
@ldoktor, that's not rude at all. If I remember correctly, the point you made was "the simpler the better" instead of "that's slow". You'd have convinced me with the later for sure.
I will send a PR for that, eventually.
Thanks