@clebergnu 2018-04-13
@ldoktor just checking if you're going to reply to https://github.com/avocado-framework/avocado/pull/2576#discussion_r181427566