@apahim 2018-04-13
@ldoktor about this PR, I will not merge it myself due to the points I have already made, but I'm not nacking it. If someone else accepts it, I'm ok with that.