@ldoktor 2018-04-13
I agree although I still don't understand what answer would satisfy you. Anyway as I said this feature is more important than the suggested simple function.

Speaking of suggestions, it'd help me not to see "NACK", but rather to see what's needed to get this merged (as the purpose of this PR is not that single commit). Anyway I sent a v2 without that commit and I hope it'll be welcomed better than this.